# go-backend-auth/docs/swagger.yaml
basePath: /
definitions:
  controllers.CreateUserInput:
    properties:
      jabatan:
        example: Kanit Reskrim
        type: string
      nama_lengkap:
        example: Budi Santoso
        type: string
      no_telp:
        example: "08123456789"
        type: string
      nrp:
        example: "87011234"
        type: string
      password:
        example: rahasia123
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Admin wajib isi Role
        example: polres
    required:
    - jabatan
    - nama_lengkap
    - no_telp
    - nrp
    - password
    - role
    type: object
  controllers.LoginInput:
    properties:
      nrp:
        example: "12345678"
        type: string
      password:
        example: password123
        type: string
    required:
    - nrp
    - password
    type: object
  controllers.RegisterInput:
    properties:
      jabatan:
        example: Anggota Sabhara
        type: string
      nama_lengkap:
        example: Budi Santoso
        type: string
      no_telp:
        description: Tambahan
        example: "08123456789"
        type: string
      nrp:
        example: "12345678"
        type: string
      password:
        example: password123
        type: string
    required:
    - jabatan
    - nama_lengkap
    - no_telp
    - nrp
    - password
    type: object
  controllers.UpdateUserInput:
    properties:
      jabatan:
        example: Kapolsek
        type: string
      nama_lengkap:
        example: Budi Santoso S.H.
        type: string
      no_telp:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        example: polres
    type: object
  models.Role:
    enum:
    - admin
    - view
    - polres
    - polsek
    type: string
    x-enum-comments:
      RoleView: Default untuk user baru
    x-enum-descriptions:
    - ""
    - Default untuk user baru
    - ""
    - ""
    x-enum-varnames:
    - RoleAdmin
    - RoleView
    - RolePolres
    - RolePolsek
host: localhost:8080
info:
  contact:
    email: admin@polri.go.id
    name: Tim IT Support
  description: API Service untuk Manajemen User (Login NRP) dan Pelaporan Data Kepolisian.
  termsOfService: http://swagger.io/terms/
  title: Backend API Polres & Polsek (Sistem NRP)
  version: "1.3"
paths:
  /admin/polres:
    get:
      description: Mengambil list semua Polres beserta detail Wilayahnya
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua Polres
      tags:
      - Master Data (Polres)
    post:
      consumes:
      - application/json
      description: Membuat data Polres dengan referensi ID Wilayah
      parameters:
      - description: Data Polres
        in: body
        name: body
        required: true
        schema:
          properties:
            kapolres:
              type: string
            nama_polres:
              type: string
            no_telp_polres:
              type: string
            wilayah_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tambah Polres Baru
      tags:
      - Master Data (Polres)
  /admin/polsek:
    get:
      description: Mengambil list semua Polsek beserta detail Polres dan Wilayahnya
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua Polsek
      tags:
      - Master Data (Polsek)
    post:
      consumes:
      - application/json
      description: Membuat data Polsek dengan referensi ID Polres dan ID Wilayah
      parameters:
      - description: Data Polsek
        in: body
        name: body
        required: true
        schema:
          properties:
            kapolsek:
              type: string
            kode:
              type: string
            nama_polsek:
              type: string
            no_telp_polsek:
              type: string
            polres_id:
              type: integer
            wilayah_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tambah Polsek Baru
      tags:
      - Master Data (Polsek)
  /admin/users:
    get:
      consumes:
      - application/json
      description: Mengambil list semua user yang terdaftar.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua User
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Membuat user baru dengan NRP dan Role tertentu.
      parameters:
      - description: Data User Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Tambah User Baru (Admin Only)
      tags:
      - admin-users
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete user berdasarkan ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Hapus User
      tags:
      - admin-users
    get:
      consumes:
      - application/json
      description: Mengambil data user berdasarkan ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Detail User
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      description: Mengupdate Nama, Jabatan, Role, atau No Telp user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit Data User (Termasuk Upgrade Role)
      tags:
      - admin-users
  /admin/wilayah:
    get:
      description: Mengambil list semua wilayah yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua Wilayah
      tags:
      - Master Data (Wilayah)
    post:
      consumes:
      - application/json
      description: Membuat data master wilayah (Kecamatan/Kabupaten)
      parameters:
      - description: Data Wilayah
        in: body
        name: body
        required: true
        schema:
          properties:
            kabupaten:
              type: string
            kecamatan:
              type: string
            latitude:
              type: number
            longitude:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Tambah Wilayah Baru
      tags:
      - Master Data (Wilayah)
  /login:
    post:
      consumes:
      - application/json
      description: Login menggunakan NRP dan Password untuk mendapatkan Token JWT.
      parameters:
      - description: Data Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login User
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Mendaftar akun baru. Default role adalah 'view'.
      parameters:
      - description: Data Registrasi
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register User Baru (Public)
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"